requirements.txt
Flask==2.3.2
pyftpdlib==1.5.6
python-nmap==0.7.1
scapy==2.4.5
pandas==2.1.1
netifaces